+++ csharp/src/Drivers/Databricks/Reader/CloudFetch/Clock.cs
@@ -17,7 +17,7 @@
 
 using System;
 
-namespace Apache.Arrow.Adbc.Drivers.Databricks.CloudFetch
+namespace Apache.Arrow.Adbc.Drivers.Databricks.Reader.CloudFetch
 {
     /// <summary>
     /// Abstraction for time operations to enable testing with controlled time.
