+++ csharp/test/Drivers/Databricks/E2E/Reader/CloudFetch/CloudFetchResultFetcherTest.cs
@@ -22,12 +22,12 @@ using System.Diagnostics;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
-using Apache.Arrow.Adbc.Drivers.Databricks.CloudFetch;
+using Apache.Arrow.Adbc.Drivers.Databricks.Reader.CloudFetch;
 using Apache.Hive.Service.Rpc.Thrift;
 using Moq;
 using Xunit;
 
-namespace Apache.Arrow.Adbc.Tests.Drivers.Databricks.CloudFetch
+namespace Apache.Arrow.Adbc.Tests.Drivers.Databricks.Reader.CloudFetch
 {
     /// <summary>
     /// Tests for CloudFetchResultFetcher
